+++ wemake_python_styleguide/visitors/presets/tokens.py
@@ -3,6 +3,9 @@
 from wemake_python_styleguide.visitors.tokenize.comments import (
     WrongCommentVisitor,
 )
+from wemake_python_styleguide.visitors.tokenize.keywords import (
+    WrongKeywordTokenVisitor,
+)
 from wemake_python_styleguide.visitors.tokenize.primitives import (
     WrongPrimitivesVisitor,
 )
@@ -10,5 +13,6 @@ from wemake_python_styleguide.visitors.tokenize.primitives import (
 #: Used to store all token related visitors to be later passed to checker:
 TOKENS_PRESET = (
     WrongCommentVisitor,
+    WrongKeywordTokenVisitor,
     WrongPrimitivesVisitor,
 )